Add revocation support.
This extracts the existing set and clear logic into a new function,
dlmalloc_revoke().  Shadow buffer pointers are hung off each segment.
To avoid looking up the segment the check is in up to three times for
each revoked chunk, store a pointer to the shadow buffer in the pad
element of the chunk.  A single lookup is required in all cases to
allow us to remove bounds from the pointer in order to access the chunk.