Actually allocate enough space for the request.
Previously a combination of accounting for too little overhead and an
improper computation of MIN_CHUNK_SIZE (negative) resulated in
pad_request() computing chunk sizes that were too small.  This resulted
in allocation bounds overlapping chunk headers.

The new compuation of MIN_CHUNK_SIZE is slightly larger than it needs to
be (a full 64 bytes of payload), but we're always paying at least a
64-byte cost so it probably doesn't make sense to allow very slightly
smaller payloads.