Zero all allocations before returning them.
This ensures that the transitive-closure of pages reachable from a
given point in the code only increases by the allocation. The actual
implementation is quite inefficent and performs unnecessicary zeroing
for newly mapped memory, values used by realloc, and (eventually) pages
that are entierly revoked and are now zeroed as a result.