Store the unbound pointer rather than the bounded pointer in pad.
Check that the pad pointer contains the CHERI_PERM_CHERIABI_VMMAP
permission.  In conjunction with a guarantee that dlmalloc never returns
pointers to the internal state (most easly done by zeroing all allocated
memory at return time), this ensures that the caller can not construct a
bogus chunk header and use it to corrupt malloc state.