Optionally unmap the interior of quarantined chunks.
When a chunk to be added to the free buffer has a least
mparams.unmap_threshold bytes of interior pages, unmap them and mark the
chunk as being unmapped.  mparams.unmap_threshold is set to PAGESIZE *
DEFAULT_UNMAP_THRESHOLD and defaults to MAX_SIZE_T / PAGESIZE disabling
unmapping.

When unmapping is enabled, maintain the invariant that either all
interior pages of a chunk are mapped or all are unmapped and it is
marked cunmapped.  This potentially allows zeroing to be limited to
the ends of the chunk.