Don't leak allocator internals.
When consolidating revoked chunks on the free list(s), clear the
allocator metadata.  When consolidating backwards, just clear the mchunk
which includes the freebuf linked list pointers and the shadow pointer.
When consolidating forward, clear as though this were a tchunk since it
might be and we always have that much space with CHERI.