Move zeroing to the free path.
I'd previously altered the consolidation path to clear allocator
metadata in the chunks that are being merged into a previous chunk so it
should be safe to rely on zeroing at free time.

Rather than enabling zeroing based on CAPREVOKE, enable it based on
ZERO_MEMORY=0/1 defaulting to on when CAPREVOKE is enabled to preserve
prior behavior.