Fix the minimum allocation alignment requirement
The nonreuse version of dlmalloc adds further bits to the low bits of the
chunk header field that is also used for the allocation size.  Therefore
higher alignment is needed.  This change was not needed on CHERI because
the default alignment is twice the pointer size, which is more than
sufficient on CHERI.